requirements.txt
attrs==25.4.0
blinker==1.9.0
brother_ql-inventree==1.3
click==8.3.1
dominate==2.9.1
Flask==3.1.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
packbits==0.6
pdf2image==1.17.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-barcode==0.16.1
pyusb==1.3.1
fonttools==4.61.1
qrcode==8.2
requests==2.32.5
visitor==0.1.3
Werkzeug==3.1.5
